Implement content type negotiation
The interrelationships between quality, precedence, and the ordering of both acceptable ranges and available types makes this algorithm a little convoluted, but it appears to produce the right results consistently.

The extensive test suite attempts to ensure full conformance with the rules of RFC9110 s12.5.1, however that spec is somewhat prone to ambiguity.